requirements.txt
fastapi
pydantic
python-dotenv
langchain-google-genai
browser-use
langchain-openai
uvicorn
requests